chore: move domain logic to customdomains package
This change moves logic related to custom domains into the customdomains
package and a couple of headers to the constants package to prevent the
following import cycle in test code:

```
  internal/gateway (test)
    ↓ imports
  internal/testenv
    ↓ imports
  internal/functions
    ↓ imports
  internal/assets
    ↓ imports
  internal/middleware
    ↓ imports
  internal/gateway (customdomains.go:16, logging.go:15)
    ↓ CYCLE!
```
