+++ server/internal/constants/http.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	HeaderProxiedResponse  = "X-Gram-Proxy-Response"
+	HeaderFilteredResponse = "X-Gram-Proxy-ResponseFiltered"
+)
