+++ server/internal/oauth/impl.go
@@ -25,6 +25,7 @@ import (
 	"github.com/speakeasy-api/gram/server/internal/attr"
 	"github.com/speakeasy-api/gram/server/internal/cache"
 	"github.com/speakeasy-api/gram/server/internal/conv"
+	"github.com/speakeasy-api/gram/server/internal/customdomains"
 	"github.com/speakeasy-api/gram/server/internal/encryption"
 	"github.com/speakeasy-api/gram/server/internal/environments"
 	"github.com/speakeasy-api/gram/server/internal/gateway"
@@ -128,7 +129,7 @@ func Attach(mux goahttp.Muxer, service *Service) {
 
 // buildFullMcpSlug builds the full MCP slug with domain prefix
 func (s *Service) buildFullMcpSlug(ctx context.Context, mcpSlug string) string {
-	if domainCtx := gateway.DomainFromContext(ctx); domainCtx != nil {
+	if domainCtx := customdomains.FromContext(ctx); domainCtx != nil {
 		return fmt.Sprintf("https://%s", domainCtx.Domain) + "/mcp/" + mcpSlug
 	}
 
@@ -136,12 +137,12 @@ func (s *Service) buildFullMcpSlug(ctx context.Context, mcpSlug string) string {
 }
 
 // loadToolsetFromMcpSlug loads a toolset from the MCP slug
-func (s *Service) loadToolsetFromMcpSlug(ctx context.Context, mcpSlug string) (*toolsets_repo.Toolset, *gateway.DomainContext, error) {
+func (s *Service) loadToolsetFromMcpSlug(ctx context.Context, mcpSlug string) (*toolsets_repo.Toolset, *customdomains.Context, error) {
 	var toolset toolsets_repo.Toolset
 	var toolsetErr error
-	var customDomainCtx *gateway.DomainContext
+	var customDomainCtx *customdomains.Context
 
-	if domainCtx := gateway.DomainFromContext(ctx); domainCtx != nil {
+	if domainCtx := customdomains.FromContext(ctx); domainCtx != nil {
 		toolset, toolsetErr = s.toolsetsRepo.GetToolsetByMcpSlugAndCustomDomain(ctx, toolsets_repo.GetToolsetByMcpSlugAndCustomDomainParams{
 			McpSlug:        conv.ToPGText(mcpSlug),
 			CustomDomainID: uuid.NullUUID{UUID: domainCtx.DomainID, Valid: true},
